chore: Pull request comments fixes
