+++ src/backend/features/payments/index.ts
@@ -1,5 +1,4 @@
 import { UserAvailableProducts } from './payments.types';
 import { getUserAvailableProducts } from './services/get-user-available-products';
-import { getUserAvailableProductsAndCallback } from './services/get-user-available-products-and-callback';
 
-export { UserAvailableProducts, getUserAvailableProducts, getUserAvailableProductsAndCallback };
+export { UserAvailableProducts, getUserAvailableProducts };
