--- src/backend/features/payments/services/are-products-equal.ts
@@ -1,12 +1,0 @@
-import { UserAvailableProducts } from '../payments.types';
-
-type Props = {
-  stored: UserAvailableProducts | undefined;
-  fetched: UserAvailableProducts;
-};
-
-export function areProductsEqual({ stored, fetched }: Props) {
-  if (!stored) return false;
-
-  return !!(stored.antivirus === fetched.antivirus && stored.backups === fetched.backups && stored.cleaner === fetched.cleaner);
-}
