+++ src/backend/features/payments/services/get-user-available-products.ts
@@ -1,10 +1,18 @@
-import { Payments } from '@internxt/sdk/dist/drive';
+import { logger } from '@/backend/core/logger/logger';
 
 import { userAvailableProductsMapper } from '../user-available-products.mapper';
+import { getPaymentsClient, PaymentsClientConfig } from './get-payments-client';
 
-type Props = { paymentsClient: Payments };
-
-export async function getUserAvailableProducts({ paymentsClient }: Props) {
-  const { featuresPerService } = await paymentsClient.getUserTier();
-  return userAvailableProductsMapper(featuresPerService);
+export async function getUserAvailableProducts({ paymentsClientConfig }: { paymentsClientConfig: PaymentsClientConfig }) {
+  try {
+    const paymentsClient = getPaymentsClient(paymentsClientConfig);
+    const userProductsInfo = await paymentsClient.getUserTier();
+    return userAvailableProductsMapper(userProductsInfo.featuresPerService);
+  } catch (error) {
+    logger.error({
+      tag: 'PRODUCTS',
+      msg: 'Failed to get user available products with error:',
+      error,
+    });
+  }
 }
