+++ src/backend/features/payments/user-available-products.mapper.test.ts
@@ -1,14 +1,14 @@
-import { DriveFeatures, Tier } from '@internxt/sdk/dist/drive/payments/types/tiers';
+import { Tier } from '@internxt/sdk/dist/drive/payments/types/tiers';
+
+import { mockProps } from '@/tests/vitest/utils.helper.test';
 
 import { userAvailableProductsMapper } from './user-available-products.mapper';
 
 describe('userAvailableProductsMapper', () => {
   it('should correctly map an object of Tier["featuresPerService"] into the proper domain object', () => {
-    const mockFeaturesPerService: Tier['featuresPerService'] = {
+    const mockFeaturesPerService: Tier['featuresPerService'] = mockProps<typeof userAvailableProductsMapper>({
       drive: {
         enabled: true,
-        maxSpaceBytes: 1,
-        workspaces: {} as DriveFeatures['workspaces'],
       },
       backups: { enabled: true },
       antivirus: { enabled: false },
@@ -16,11 +16,11 @@ describe('userAvailableProductsMapper', () => {
       mail: { enabled: false, addressesPerUser: 0 },
       vpn: { enabled: false, featureId: '123' },
       cleaner: { enabled: true },
-    };
+    });
 
     const result = userAvailableProductsMapper(mockFeaturesPerService);
 
-    expect(result).toEqual({
+    expect(result).toStrictEqual({
       backups: true,
       antivirus: false,
       cleaner: true,
@@ -28,15 +28,15 @@ describe('userAvailableProductsMapper', () => {
   });
 
   it('should correctly map into the proper domain object even though we recieve incorrect properties', () => {
-    const mockFeaturesPerService = {
+    const mockFeaturesPerService: Tier['featuresPerService'] = mockProps<typeof userAvailableProductsMapper>({
       vpn: { enabled: false, featureId: '123' },
       antivirus: { enabled: null } as unknown as Tier['featuresPerService']['antivirus'],
       backups: { enabled: true },
-    } as unknown as Tier['featuresPerService'];
+    });
 
     const result = userAvailableProductsMapper(mockFeaturesPerService);
 
-    expect(result).toEqual({
+    expect(result).toStrictEqual({
       backups: true,
       antivirus: false,
       cleaner: false,
