+++ src/backend/features/payments/user-available-products.mapper.ts
@@ -1,8 +1,6 @@
 import { Tier } from '@internxt/sdk/dist/drive/payments/types/tiers';
 
-import { UserAvailableProducts } from './payments.types';
-
-export function userAvailableProductsMapper(featuresPerService: Tier['featuresPerService']): UserAvailableProducts {
+export function userAvailableProductsMapper(featuresPerService: Tier['featuresPerService']) {
   return {
     backups: !!featuresPerService['backups']?.enabled,
     antivirus: !!featuresPerService['antivirus']?.enabled,
